+++ CHANGELOG.md
@@ -17,6 +17,8 @@ A dependency update to an incompatible version is considered a breaking change.
 
 ## Unreleased
 
+## v0.1.2 - 2024-12-14
+
 ### Added
 
 - `Attr::set`